move parseFullXMLDocs to File.swift
